Move all classes out of control to respective packages
