+++ common/src/main/java/us/ilite/common/lib/control/AController.java
@@ -1,10 +1,11 @@
-package control;
+package us.ilite.common.lib.control;
 
 import com.team254.lib.geometry.Pose2d;
 import com.team254.lib.geometry.Pose2dWithCurvature;
 import com.team254.lib.physics.DifferentialDrive;
 import com.team254.lib.trajectory.TrajectoryIterator;
 import com.team254.lib.trajectory.timing.TimedState;
+import us.ilite.common.lib.DriveOutput;
 
 public abstract class AController {
