+++ common/src/main/java/us/ilite/common/lib/odometry/Kinematics.java
@@ -1,6 +1,6 @@
-package odometry;
+package us.ilite.common.lib.odometry;
 
-import profiles.RobotProfile;
+import us.ilite.common.lib.RobotProfile;
 import com.team254.lib.geometry.Pose2d;
 import com.team254.lib.geometry.Rotation2d;
 import com.team254.lib.geometry.Twist2d;
