--- control/src/main/java/paths/autos/FarScaleAuto.java
@@ -1,5 +1,0 @@
-package paths.autos;
-
-public class FarScaleAuto {
-
-}
