+++ robot/src/main/java/us/ilite/robot/Robot.java
@@ -3,11 +3,11 @@ package us.ilite.robot;
 import com.flybotix.hfr.util.log.ELevel;
 import com.flybotix.hfr.util.log.ILog;
 import com.flybotix.hfr.util.log.Logger;
-import control.DriveController;
+import us.ilite.common.lib.control.DriveController;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.Timer;
-import paths.TrajectoryGenerator;
-import paths.autos.TestAuto;
+import us.ilite.common.lib.TrajectoryGenerator;
+import us.ilite.robot.auto.paths.TestAuto;
 import us.ilite.common.config.SystemSettings;
 import com.team254.lib.geometry.Pose2dWithCurvature;
 import com.team254.lib.trajectory.Trajectory;
