+++ robot/src/main/java/us/ilite/robot/auto/paths/FarScaleAuto.java
@@ -1,0 +1,5 @@
+package us.ilite.robot.auto.paths;
+
+public class FarScaleAuto {
+
+}
