+++ robot/src/main/java/us/ilite/robot/auto/paths/NearScaleAuto.java
@@ -1,4 +1,4 @@
-package paths.autos;
+package us.ilite.robot.auto.paths;
 
 import com.team254.lib.geometry.Pose2d;
 import com.team254.lib.geometry.Rotation2d;
